+++ packages/flutter_tools/test/crash_reporting_test.dart
@@ -13,7 +13,7 @@ import 'package:http/http.dart';
 import 'package:http/testing.dart';
 import 'package:test/test.dart';
 
-import 'package:flutter_tools/executable.dart' as tools;
+import 'package:flutter_tools/runner.dart' as tools;
 import 'package:flutter_tools/src/base/context.dart';
 import 'package:flutter_tools/src/base/io.dart';
 import 'package:flutter_tools/src/base/logger.dart';
